refactor - CompanySettingsLogoController uses Service/Repository pattern

+++ app/Services/CompanySettingsLogo/CompanySettingsLogoService.php
@@ -3,6 +3,7 @@
 namespace App\Services\CompanySettingsLogo;
 
 use App\Repositories\Company\CompanyRepository;
+use Exception;
 use Illuminate\Support\Facades\Storage;
 
 class CompanySettingsLogoService{
@@ -10,6 +11,12 @@ class CompanySettingsLogoService{
 	public function __construct(private CompanyRepository $company_repository){
 	}
 
+	/**
+	 * fetch function
+	 *
+	 * @param integer $company_id
+	 * @return string
+	 */
 	public function fetch(int $company_id) : string {
 
 		$path = 'logos/'.$company_id;
@@ -28,34 +35,54 @@ class CompanySettingsLogoService{
 
 	}
 
-	public function updateCompanyLogo(array $data){
+	/**
+	 * update function
+	 *
+	 * @param array $data
+	 * @return boolean
+	 */
+	public function update(array $data) : bool {
+		
+		$company_id = $data['company_id'];
+		$file = $data['logo'];
+		$path = 'logos/'.$company_id;
+		
+		$extension = $file->getClientOriginalExtension();
+		$filename = md5(time().'_'.$file->getClientOriginalName()).'.'.$extension;
+		
+		Storage::disk('public')->deleteDirectory($path);
+		Storage::disk('public')->makeDirectory($path);
 
-		//if($request->hasFile('logo')){
-			$company_id = $data['company_id'];
-			$file = $data['logo'];
+		Storage::disk('public')->putFileAs($path, $file, $filename);
+
+		$company = $this->company_repository->fetchDefaultById($company_id);
+		return $this->company_repository->updateCompanyLogoByObj($filename, $company);
+		
+
+	}
+
+	/**
+	 * remove function
+	 *
+	 * @param integer $company_id
+	 * @return boolean
+	 */
+	public function remove(int $company_id) : bool {
+		
+		try{
+
 			$path = 'logos/'.$company_id;
-			
-			$extension = $file->getClientOriginalExtension();
-			$filename = md5(time().'_'.$file->getClientOriginalName()).'.'.$extension;
-			
-			Storage::disk('public')->deleteDirectory($path);
-			Storage::disk('public')->makeDirectory($path);
 
-			Storage::disk('public')->putFileAs($path, $file, $filename);
+			if(Storage::disk('public')->exists($path)){
+				Storage::disk('public')->deleteDirectory($path);
+			}
 
 			$company = $this->company_repository->fetchDefaultById($company_id);
-			$this->company_repository->updateCompanyLogoByObj($filename, $company);
-			
-			// $company = General::fetchDefaultCompanyById($company_id);
-			// $company->logo = $filename;
+			return $this->company_repository->updateCompanyLogoByObj('', $company);
 
-			// if($company->save()){
-			// 	return response(['message' => 'Logo uploaded successfully', 'validity' => 'upload_success'], 200);
-			// }
-
-			
-		//}
-
+		}catch(Exception $e){
+			throw new Exception('failed to remove logo');
+		}
 	}
 
 }
